FT_PRINTF(maybe fixed %p need to check on school mac bc of M2)

+++ ft_printf/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marbaron <marbaron@student.42.fr>          +#+  +:+       +#+        */
+/*   By: margueritebaronbeliveau <margueritebaro    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/04 10:57:02 by marbaron          #+#    #+#             */
-/*   Updated: 2024/03/12 12:52:35 by marbaron         ###   ########.fr       */
+/*   Updated: 2024/03/12 16:31:27 by margueriteb      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ int		ft_strlen(char *str);
 size_t	ft_putchar_pf(char c);
 size_t	ft_putstr_pf(char *s);
 size_t	ft_putnbr_pf(long long n, int not_unsigned);
-size_t	ft_puthex(unsigned int n, int letter);
+size_t	ft_puthex(unsigned long n, int letter);
 size_t	ft_pointer(void *ptr);
 
 #endif
